+++ firmware/lib/modules/transmission.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-uint8_t calculate_checksum(void *buffer, uint8_t len) {
+uint8_t calculate_checksum(void *buffer, uint16_t len) {
     uint8_t checksum = 0;
     char *bytes = (char *)buffer;
     for (int i = 0; i < len; i++) {
